Fix 64-bit UB in BinaryenLiteral (#3136)
The maximum storage size of type ids changed with the introduction of tuples, with their type id being the memory address of the canonicalized type vector. This change aligns the storage type of type ids in `BinaryenLiteral` with `Literal` by changing the type from `int32_t` to `uintptr_t`.

+++ src/binaryen-c.h
@@ -223,7 +223,7 @@ BINARYEN_API void BinaryenModuleDispose(BinaryenModuleRef module);
 // Literals. These are passed by value.
 
 struct BinaryenLiteral {
-  int32_t type;
+  uintptr_t type;
   union {
     int32_t i32;
     int64_t i64;
@@ -231,6 +231,7 @@ struct BinaryenLiteral {
     double f64;
     uint8_t v128[16];
     const char* func;
+    // TODO: exn
   };
 };
